Add files via upload
The first project of Frontend Web Development; a letter to the college dean to conduct a coding camp for 30 days by Devsnest. It's the basic outline of the letter made using only HTML. In this, I have used tags such as <hr> and <table>---</table>. I have given different backgrounds colors and fonts.